Return true or false if record is deleted

+++ app/Http/Controllers/Api/V1/GroupMemberController.php
@@ -92,15 +92,29 @@ class GroupMemberController extends Controller
         $group = Group::where('uuid', $request->group_id)->first();
         $user = User::where('uuid', $request->receiver_id)->first();
 
+        $deleted_1 = false;
+        $deleted_2 = false;
         try {
             if($group->status != 0) { //if group in private then delete its invitation also
-                GroupInvitation::where('group_id', $group->id)->where('invitee_user_id', $user->id)->delete();
+                $group_invitation = GroupInvitation::where('group_id', $group->id)->where('invitee_user_id', $user->id)->first();
+                if($group_invitation){
+                    if($group_invitation->status == 1){
+                        $group_invitation->delete();
+                        $deleted_1 = true;
+                    }
+                }
             }
 
-            $group_member = GroupMember::where('user_id', $user->id)->where('group_id', $group->id)->firstOrFail();
-            $group_member->delete();
+            $group_member = GroupMember::where('user_id', $user->id)->where('group_id', $group->id)->first();
+            if($group_member){
+                $group_member->delete();
+                $deleted_2 = true;
+            }
+            if($deleted_1 || $deleted_2){
+                return "true";
+            }
+            return "false";
 
-            return ApiResponse::success(new GroupMemberResource($group_member), 200);
         } catch (\Exception $e) {
             return ApiResponse::error(trans('response.not_found'), 404);
         }
